… .babelrc
… .gitignore
… CHANGELOG.md
… LICENSE
… README.md
… bench/index.js
… lexer.js
… lib/lexer.wasm
… package.json
… src/lexer.js
… test/_unit.js
… test/integration.js
… test/samples/d3.js
… test/samples/magic-string.js
… test/samples/rollup.js