This converts the project into C with Wasm, resulting in a self-contained 10KiB .js file (including the Wasm binary), which is 4KiB minified.

Resolves: #8, #4, #3.

Tests passing. The API remains identical such that this is a patch release.

The benchmarks show a 6x improvement on cold start, which is the major justification here for Wasm conversion, given that this project needs to run on bootstrap so cannot rely on speeding up after the first X megabytes of source transform runs - it needs to be fast from the start.

Once v8 optimizations kick in performance is roughly the same, although further optimization is likely possible on top of this work.

Before conversion:

```
Cold Run, All Samples
bench/samples/*.js (2195 KiB)
> 179ms

Warm Runs (average of 25 runs)
bench/samples/d3.js (508 KiB)
> 6.84ms
bench/samples/d3.min.js (274 KiB)
> 3.16ms
bench/samples/magic-string.js (35 KiB)
> 0.6ms
bench/samples/magic-string.min.js (20 KiB)
> 0ms
bench/samples/rollup.js (929 KiB)
> 8.64ms
bench/samples/rollup.min.js (429 KiB)
> 5.12ms

Warm Runs, All Samples (average of 25 runs)
bench/samples/*.js (2195 KiB)
> 24.24ms
```

After conversion:

```
Cold Run, All Samples
bench/samples/*.js (2195 KiB)
> 33ms

Warm Runs (average of 25 runs)
test/samples/d3.js (508 KiB)
> 5.44ms
test/samples/d3.min.js (274 KiB)
> 3ms
test/samples/magic-string.js (35 KiB)
> 0.16ms
test/samples/magic-string.min.js (20 KiB)
> 0ms
test/samples/rollup.js (929 KiB)
> 9.2ms
test/samples/rollup.min.js (429 KiB)
> 5.04ms

Warm Runs, All Samples (average of 25 runs)
bench/samples/*.js (2195 KiB)
> 21.84ms
```