@LarsDenBakker 2019-08-21
This is really interesting. What does this do in terms of browser and node support?