@guybedford 2019-08-21
All environments that support ES modules already support Web Assembly, so this works in all the same environments.