@jimmywarting 2023-07-03
currently the module is inlined as base64... and you also must call init to wait for the wasm to finish...
would it be better to use top level await + import.meta.resolve instead?

```js
const url = import.meta.resolve('./simple.wasm')
// alt.
const url = new URL('./simple.wasm', import.meta.url).toString()

const mod = await WebAssembly.instantiateStreaming(fetch(url), { ... })
export default instance.exports
```